added noUpdate function

+++ ftbContainer/files/docker-entrypoint.sh
@@ -19,19 +19,21 @@ unzip  /tmp/Files.zip -d /tmp/Server
 echo "comapring JSONS"
 
 echo "already running Server found"
-if ! cmp --silent /opt/Server/version.json /tmp/Server/version.json ; then
-    echo "Json files differ, Update"
-    # Update Logic
-    # Overwrite every File/Dir in the Server directory if the corresponding file/dir in the archive exists
-    for i in $(ls /tmp/Server); do
-        if [ -e /opt/Server/${i} ]; then
-            echo "File ${i} in opt/Server found"
-            rm -r $SERVER_VOL/${i}
-            echo "Deleted File/Dir ${i}"
-        fi
-        cp -r /tmp/Server/${i} $SERVER_VOL/${i}
-        echo "Copied ${i} to Server Dir"
-    done
+if ! $NO_UPDATE; then 
+    if ! cmp --silent /opt/Server/version.json /tmp/Server/version.json ; then
+        echo "Json files differ, Update"
+        # Update Logic
+        # Overwrite every File/Dir in the Server directory if the corresponding file/dir in the archive exists
+        for i in $(ls /tmp/Server); do
+            if [ -e /opt/Server/${i} ]; then
+                echo "File ${i} in opt/Server found"
+                rm -r $SERVER_VOL/${i}
+                echo "Deleted File/Dir ${i}"
+            fi
+            cp -r /tmp/Server/${i} $SERVER_VOL/${i}
+            echo "Copied ${i} to Server Dir"
+        done
+    fi
 fi
 
 rm -r /tmp/Files.zip
